rename: my-profile fix
