+++ src/app/myprofile/_components/index.ts
@@ -1,2 +1,2 @@
-export { default as MyProfile } from "./my-profile/MyProfile";
+export { default as MyProfile } from "./my-profile/my-profile";
 export { default as ReviewItem } from "./review-item/review-item";
